requirements.txt
grpcio==1.46.1
numpy==1.19.2
pandas==1.1.5
protobuf==3.20.0
